+++ app/src/main/java/jetpack/tutorial/cryptoapp/features/crypto/data/remote/dto/CryptoListingDto.kt
@@ -7,5 +7,5 @@ data class CryptoListingDto(
     @field:Json(name = "symbol") val symbol: String,
     @field:Json(name = "name") val name: String,
     @field:Json(name = "image") val image: String,
-    @field:Json(name = "current_price") val currentPrice: Int,
+    @field:Json(name = "current_price") val currentPrice: Double,
 )
