+++ app/src/main/java/jetpack/tutorial/cryptoapp/features/crypto/domain/model/CryptoListingModel.kt
@@ -5,5 +5,5 @@ data class CryptoListingModel(
     val symbol: String,
     val name: String,
     val image: String,
-    val currentPrice: Int,
+    val currentPrice: Double,
 )
